requirements.txt
# 基础科学计算库
numpy>=1.24,<2.0  # PyTorch兼容性要求
pandas>=2.0

# 机器学习框架
scikit-learn>=1.3

# 深度学习框架 (PyTorch)
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.13.0

# 图像处理
opencv-contrib-python>=4.8,<4.10  # 兼容 NumPy 1.x
scikit-image>=0.21
Pillow>=9.5.0

# 可视化
matplotlib>=3.7
seaborn>=0.12

# 工具库
tqdm>=4.66
joblib>=1.3

# Web UI
gradio>=4.0